refactor: move admin rights helpers to commands module
